Revert "Avatar V2 prototype (#236)"
