+++ pallets/ajuna-awesome-avatars/benchmarking/src/lib.rs
@@ -120,12 +120,7 @@ fn create_avatars<T: Config>(name: &'static str, n: u32) -> Result<(), &'static 
 	for _ in 0..n {
 		AAvatars::<T>::mint(
 			RawOrigin::Signed(player.clone()).into(),
-			MintOption {
-				mint_type: MintType::Free,
-				count: MintPackSize::One,
-				mint_pack: PackType::Material,
-				mint_version: AvatarVersion::V1,
-			},
+			MintOption { mint_type: MintType::Free, count: MintPackSize::One },
 		)?;
 		Accounts::<T>::mutate(&player, |account| account.stats.mint.last = Zero::zero());
 	}
@@ -184,8 +179,7 @@ benchmarks! {
 		let caller = account::<T>(name);
 		Accounts::<T>::mutate(&caller, |account| account.free_mints = MintCount::MAX);
 
-		let mint_option = MintOption { mint_type: MintType::Free, count: MintPackSize::Six,
-			mint_pack: PackType::Material, mint_version: AvatarVersion::V1 };
+		let mint_option = MintOption { mint_type: MintType::Free, count: MintPackSize::Six };
 	}: mint(RawOrigin::Signed(caller.clone()), mint_option)
 	verify {
 		let n = n as usize;
@@ -202,8 +196,7 @@ benchmarks! {
 		let mint_fee = GlobalConfigs::<T>::get().mint.fees.fee_for(&MintPackSize::Six);
 		CurrencyOf::<T>::make_free_balance_be(&caller, mint_fee);
 
-		let mint_option = MintOption { mint_type: MintType::Normal, count: MintPackSize::Six,
-			mint_pack: PackType::Material, mint_version: AvatarVersion::V1 };
+		let mint_option = MintOption { mint_type: MintType::Normal, count: MintPackSize::Six };
 	}: mint(RawOrigin::Signed(caller.clone()), mint_option)
 	verify {
 		let n = n as usize;
@@ -233,7 +226,7 @@ benchmarks! {
 				count
 			}
 		);
-		assert_last_event::<T>(Event::AvatarsForged { avatar_ids: vec![(avatar_id, upgraded_components)] })
+		assert_last_event::<T>(Event::AvatarForged { avatar_id, upgraded_components })
 	}
 
 	transfer_avatar_normal {
