+++ pallets/ajuna-awesome-avatars/src/types/account.rs
@@ -18,15 +18,20 @@ use super::{MintCount, SeasonId};
 use frame_support::pallet_prelude::*;
 use sp_runtime::{traits::Get, BoundedBTreeSet};
 
-#[derive(Encode, Decode, MaxEncodedLen, TypeInfo, Clone, Debug, Default, PartialEq)]
+#[derive(Encode, Decode, MaxEncodedLen, TypeInfo, Clone, Debug, PartialEq)]
 pub enum StorageTier {
-	#[default]
 	One = 25,
 	Two = 50,
 	Three = 75,
 	Four = 100,
 	Five = 150,
 	Max = 200,
+}
+
+impl Default for StorageTier {
+	fn default() -> Self {
+		Self::One
+	}
 }
 
 impl StorageTier {
