+++ pallets/ajuna-awesome-avatars/src/types/config.rs
@@ -14,16 +14,20 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-use crate::types::AvatarVersion;
 use frame_support::pallet_prelude::*;
 
 /// Number of avatars to be minted.
-#[derive(Encode, Decode, MaxEncodedLen, TypeInfo, Clone, Debug, Default, PartialEq)]
+#[derive(Encode, Decode, MaxEncodedLen, TypeInfo, Clone, Copy, Debug, Eq, PartialEq)]
 pub enum MintPackSize {
-	#[default]
 	One = 1,
 	Three = 3,
 	Six = 6,
+}
+
+impl Default for MintPackSize {
+	fn default() -> Self {
+		MintPackSize::One
+	}
 }
 
 impl MintPackSize {
@@ -33,7 +37,7 @@ impl MintPackSize {
 }
 
 /// Minting fee per pack of avatars.
-#[derive(Encode, Decode, MaxEncodedLen, TypeInfo, Clone, Debug, Default, PartialEq)]
+#[derive(Encode, Decode, MaxEncodedLen, TypeInfo, Clone, Copy, Debug, Default, PartialEq)]
 pub struct MintFees<Balance> {
 	pub one: Balance,
 	pub three: Balance,
@@ -50,32 +54,25 @@ impl<Balance> MintFees<Balance> {
 	}
 }
 
-#[derive(Encode, Decode, MaxEncodedLen, TypeInfo, Clone, Debug, Default, PartialEq)]
+#[derive(Encode, Decode, MaxEncodedLen, TypeInfo, Clone, Debug, Eq, PartialEq)]
 pub enum MintType {
 	/// Mint using free mint credits.
-	#[default]
 	Free,
 	/// Normal minting consuming currency.
 	Normal,
 }
 
-#[derive(Encode, Decode, MaxEncodedLen, TypeInfo, Clone, Debug, Default, PartialEq)]
-pub enum PackType {
-	#[default]
-	Material = 1,
-	Equipment = 2,
-	Special = 3,
+impl Default for MintType {
+	fn default() -> Self {
+		MintType::Free
+	}
 }
 
 /// Minting options
-#[derive(Encode, Decode, MaxEncodedLen, TypeInfo, Clone, Debug, Default, PartialEq)]
+#[derive(Encode, Decode, MaxEncodedLen, TypeInfo, Clone, Debug, Default, Eq, PartialEq)]
 pub struct MintOption {
 	/// Type of minting.
 	pub mint_type: MintType,
-	/// Type of pack to mint
-	pub mint_pack: PackType,
-	/// Avatar version to mint
-	pub mint_version: AvatarVersion,
 	/// Number of avatars to mint.
 	pub count: MintPackSize,
 }
